Move `DatabaseInterface::initRelationParamsCache` method to Relation class
