+++ test/classes/ConfigStorage/RelationTest.php
@@ -1511,4 +1511,449 @@ class RelationTest extends AbstractTestCase
             $this->relation->getDefaultPmaTableNames(['pma__export_templates' => 'db_exporttemplates_pma'])
         );
     }
+
+    public function testInitRelationParamsCacheDefaultDbNameDbDoesNotExist(): void
+    {
+        parent::setGlobalDbi();
+
+        $GLOBALS['db'] = '';
+        $GLOBALS['server'] = 0;
+
+        $this->dummyDbi->removeDefaultResults();
+        $this->dummyDbi->addResult('SHOW TABLES FROM `phpmyadmin`;', false);
+
+        $relation = new Relation($this->dbi);
+        $relation->initRelationParamsCache();
+
+        $this->assertAllQueriesConsumed();
+    }
+
+    public function testInitRelationParamsCacheDefaultDbNameDbExistsServerZero(): void
+    {
+        parent::setGlobalDbi();
+
+        $GLOBALS['db'] = '';
+        $GLOBALS['server'] = 0;
+        $GLOBALS['cfg']['Server'] = [];
+
+        $this->dummyDbi->removeDefaultResults();
+        $this->dummyDbi->addResult(
+            'SHOW TABLES FROM `phpmyadmin`;',
+            [
+                ['pma__userconfig'],
+            ],
+            ['Tables_in_phpmyadmin']
+        );
+
+        $relation = new Relation($this->dbi);
+        $relation->initRelationParamsCache();
+
+        $this->assertArrayHasKey('relation', $_SESSION, 'The cache is expected to be filled');
+
+        // Should all be false for server = 0
+        $this->assertSame([
+            'version' => $_SESSION['relation'][$GLOBALS['server']]['version'],
+            'relwork' => false,
+            'displaywork' => false,
+            'bookmarkwork' => false,
+            'pdfwork' => false,
+            'commwork' => false,
+            'mimework' => false,
+            'historywork' => false,
+            'recentwork' => false,
+            'favoritework' => false,
+            'uiprefswork' => false,
+            'trackingwork' => false,
+            'userconfigwork' => false,
+            'menuswork' => false,
+            'navwork' => false,
+            'savedsearcheswork' => false,
+            'centralcolumnswork' => false,
+            'designersettingswork' => false,
+            'exporttemplateswork' => false,
+            'allworks' => false,
+            'user' => null,
+            'db' => null,
+        ], $_SESSION['relation'][$GLOBALS['server']]);
+
+        $this->assertEquals([
+            'userconfig' => 'pma__userconfig',
+            'pmadb' => false,// This is the expected value for server = 0
+        ], $GLOBALS['cfg']['Server']);
+        $this->assertAllQueriesConsumed();
+    }
+
+    public function testInitRelationParamsCacheDefaultDbNameDbExistsFirstServer(): void
+    {
+        parent::setGlobalDbi();
+
+        $GLOBALS['db'] = '';
+        $GLOBALS['server'] = 1;
+        $GLOBALS['cfg']['Server'] = [];
+        $GLOBALS['cfg']['Server']['user'] = '';
+        $GLOBALS['cfg']['Server']['pmadb'] = '';
+        $GLOBALS['cfg']['Server']['bookmarktable'] = '';
+        $GLOBALS['cfg']['Server']['relation'] = '';
+        $GLOBALS['cfg']['Server']['table_info'] = '';
+        $GLOBALS['cfg']['Server']['table_coords'] = '';
+        $GLOBALS['cfg']['Server']['column_info'] = '';
+        $GLOBALS['cfg']['Server']['pdf_pages'] = '';
+        $GLOBALS['cfg']['Server']['history'] = '';
+        $GLOBALS['cfg']['Server']['recent'] = '';
+        $GLOBALS['cfg']['Server']['favorite'] = '';
+        $GLOBALS['cfg']['Server']['table_uiprefs'] = '';
+        $GLOBALS['cfg']['Server']['tracking'] = '';
+        $GLOBALS['cfg']['Server']['userconfig'] = '';
+        $GLOBALS['cfg']['Server']['users'] = '';
+        $GLOBALS['cfg']['Server']['usergroups'] = '';
+        $GLOBALS['cfg']['Server']['navigationhiding'] = '';
+        $GLOBALS['cfg']['Server']['savedsearches'] = '';
+        $GLOBALS['cfg']['Server']['central_columns'] = '';
+        $GLOBALS['cfg']['Server']['designer_settings'] = '';
+        $GLOBALS['cfg']['Server']['export_templates'] = '';
+
+        $this->dummyDbi->removeDefaultResults();
+        $this->dummyDbi->addResult(
+            'SHOW TABLES FROM `phpmyadmin`;',
+            [
+                ['pma__userconfig'],
+            ],
+            ['Tables_in_phpmyadmin']
+        );
+
+        $this->dummyDbi->addResult(
+            'SHOW TABLES FROM `phpmyadmin`',
+            [
+                ['pma__userconfig'],
+            ],
+            ['Tables_in_phpmyadmin']
+        );
+
+        $this->dummyDbi->addResult(
+            'SELECT NULL FROM `pma__userconfig` LIMIT 0',
+            [
+                ['NULL'],
+            ],
+            ['NULL']
+        );
+
+        $this->dummyDbi->addSelectDb('phpmyadmin');
+        $relation = new Relation($this->dbi);
+        $relation->initRelationParamsCache();
+        $this->assertAllSelectsConsumed();
+
+        $this->assertArrayHasKey('relation', $_SESSION, 'The cache is expected to be filled');
+
+        // Should all be false for server = 0
+        $this->assertSame([
+            'version' => $_SESSION['relation'][$GLOBALS['server']]['version'],
+            'relwork' => false,
+            'displaywork' => false,
+            'bookmarkwork' => false,
+            'pdfwork' => false,
+            'commwork' => false,
+            'mimework' => false,
+            'historywork' => false,
+            'recentwork' => false,
+            'favoritework' => false,
+            'uiprefswork' => false,
+            'trackingwork' => false,
+            'userconfigwork' => true,
+            'menuswork' => false,
+            'navwork' => false,
+            'savedsearcheswork' => false,
+            'centralcolumnswork' => false,
+            'designersettingswork' => false,
+            'exporttemplateswork' => false,
+            'allworks' => false,
+            'user' => '',
+            'db' => 'phpmyadmin',
+            'userconfig' => 'pma__userconfig',
+        ], $_SESSION['relation'][$GLOBALS['server']]);
+
+        $this->assertSame([
+            'user' => '',
+            'pmadb' => 'phpmyadmin',
+            'bookmarktable' => '',
+            'relation' => '',
+            'table_info' => '',
+            'table_coords' => '',
+            'column_info' => '',
+            'pdf_pages' => '',
+            'history' => '',
+            'recent' => '',
+            'favorite' => '',
+            'table_uiprefs' => '',
+            'tracking' => '',
+            'userconfig' => 'pma__userconfig',
+            'users' => '',
+            'usergroups' => '',
+            'navigationhiding' => '',
+            'savedsearches' => '',
+            'central_columns' => '',
+            'designer_settings' => '',
+            'export_templates' => '',
+        ], $GLOBALS['cfg']['Server']);
+
+        $this->assertAllQueriesConsumed();
+    }
+
+    public function testInitRelationParamsCacheDefaultDbNameDbExistsFirstServerNotWorkingTable(): void
+    {
+        parent::setGlobalDbi();
+
+        $GLOBALS['db'] = '';
+        $GLOBALS['server'] = 1;
+        $GLOBALS['cfg']['Server'] = [];
+        $GLOBALS['cfg']['Server']['user'] = '';
+        $GLOBALS['cfg']['Server']['pmadb'] = '';
+        $GLOBALS['cfg']['Server']['bookmarktable'] = '';
+        $GLOBALS['cfg']['Server']['relation'] = '';
+        $GLOBALS['cfg']['Server']['table_info'] = '';
+        $GLOBALS['cfg']['Server']['table_coords'] = '';
+        $GLOBALS['cfg']['Server']['column_info'] = '';
+        $GLOBALS['cfg']['Server']['pdf_pages'] = '';
+        $GLOBALS['cfg']['Server']['history'] = '';
+        $GLOBALS['cfg']['Server']['recent'] = '';
+        $GLOBALS['cfg']['Server']['favorite'] = '';
+        $GLOBALS['cfg']['Server']['table_uiprefs'] = '';
+        $GLOBALS['cfg']['Server']['tracking'] = '';
+        $GLOBALS['cfg']['Server']['userconfig'] = '';
+        $GLOBALS['cfg']['Server']['users'] = '';
+        $GLOBALS['cfg']['Server']['usergroups'] = '';
+        $GLOBALS['cfg']['Server']['navigationhiding'] = '';
+        $GLOBALS['cfg']['Server']['savedsearches'] = '';
+        $GLOBALS['cfg']['Server']['central_columns'] = '';
+        $GLOBALS['cfg']['Server']['designer_settings'] = '';
+        $GLOBALS['cfg']['Server']['export_templates'] = '';
+
+        $this->dummyDbi->removeDefaultResults();
+        $this->dummyDbi->addResult(
+            'SHOW TABLES FROM `phpmyadmin`;',
+            [
+                ['pma__userconfig'],
+            ],
+            ['Tables_in_phpmyadmin']
+        );
+
+        $this->dummyDbi->addResult(
+            'SHOW TABLES FROM `phpmyadmin`',
+            [
+                ['pma__userconfig'],
+            ],
+            ['Tables_in_phpmyadmin']
+        );
+
+        $this->dummyDbi->addResult('SELECT NULL FROM `pma__userconfig` LIMIT 0', false);
+
+        $this->dummyDbi->addSelectDb('phpmyadmin');
+        $relation = new Relation($this->dbi);
+        $relation->initRelationParamsCache();
+        $this->assertAllSelectsConsumed();
+
+        $this->assertArrayHasKey('relation', $_SESSION, 'The cache is expected to be filled');
+
+        $this->assertSame([
+            'version' => $_SESSION['relation'][$GLOBALS['server']]['version'],
+            'relwork' => false,
+            'displaywork' => false,
+            'bookmarkwork' => false,
+            'pdfwork' => false,
+            'commwork' => false,
+            'mimework' => false,
+            'historywork' => false,
+            'recentwork' => false,
+            'favoritework' => false,
+            'uiprefswork' => false,
+            'trackingwork' => false,
+            'userconfigwork' => false,// Expected value for not working table
+            'menuswork' => false,
+            'navwork' => false,
+            'savedsearcheswork' => false,
+            'centralcolumnswork' => false,
+            'designersettingswork' => false,
+            'exporttemplateswork' => false,
+            'allworks' => false,
+            'user' => '',
+            'db' => 'phpmyadmin',
+            'userconfig' => 'pma__userconfig',
+        ], $_SESSION['relation'][$GLOBALS['server']]);
+
+        $this->assertSame([
+            'user' => '',
+            'pmadb' => 'phpmyadmin',
+            'bookmarktable' => '',
+            'relation' => '',
+            'table_info' => '',
+            'table_coords' => '',
+            'column_info' => '',
+            'pdf_pages' => '',
+            'history' => '',
+            'recent' => '',
+            'favorite' => '',
+            'table_uiprefs' => '',
+            'tracking' => '',
+            'userconfig' => 'pma__userconfig',
+            'users' => '',
+            'usergroups' => '',
+            'navigationhiding' => '',
+            'savedsearches' => '',
+            'central_columns' => '',
+            'designer_settings' => '',
+            'export_templates' => '',
+        ], $GLOBALS['cfg']['Server']);
+
+        $this->assertAllQueriesConsumed();
+    }
+
+    public function testInitRelationParamsCacheDefaultDbNameDbExistsFirstServerOverride(): void
+    {
+        parent::setGlobalDbi();
+
+        $GLOBALS['db'] = '';
+        $GLOBALS['server'] = 1;
+        $GLOBALS['cfg']['Server'] = [];
+        $GLOBALS['cfg']['Server']['user'] = '';
+        $GLOBALS['cfg']['Server']['pmadb'] = 'PMA-storage';
+        $GLOBALS['cfg']['Server']['bookmarktable'] = '';
+        $GLOBALS['cfg']['Server']['relation'] = '';
+        $GLOBALS['cfg']['Server']['table_info'] = '';
+        $GLOBALS['cfg']['Server']['table_coords'] = '';
+        $GLOBALS['cfg']['Server']['column_info'] = '';
+        $GLOBALS['cfg']['Server']['pdf_pages'] = '';
+        $GLOBALS['cfg']['Server']['history'] = '';
+        $GLOBALS['cfg']['Server']['recent'] = '';
+        $GLOBALS['cfg']['Server']['favorite'] = '';
+        $GLOBALS['cfg']['Server']['table_uiprefs'] = '';
+        $GLOBALS['cfg']['Server']['tracking'] = '';
+        $GLOBALS['cfg']['Server']['userconfig'] = 'pma__userconfig_custom';
+        $GLOBALS['cfg']['Server']['users'] = '';
+        $GLOBALS['cfg']['Server']['usergroups'] = '';
+        $GLOBALS['cfg']['Server']['navigationhiding'] = '';
+        $GLOBALS['cfg']['Server']['savedsearches'] = '';
+        $GLOBALS['cfg']['Server']['central_columns'] = '';
+        $GLOBALS['cfg']['Server']['designer_settings'] = '';
+        $GLOBALS['cfg']['Server']['export_templates'] = '';
+
+        $this->dummyDbi->removeDefaultResults();
+        $this->dummyDbi->addResult(
+            'SHOW TABLES FROM `PMA-storage`;',
+            [
+                [
+                    'pma__userconfig_custom',
+                    'pma__usergroups',
+                ],
+            ],
+            ['Tables_in_PMA-storage']
+        );
+
+        $this->dummyDbi->addResult(
+            'SHOW TABLES FROM `PMA-storage`',
+            [
+                [
+                    'pma__userconfig_custom',
+                    'pma__usergroups',
+                ],
+            ],
+            ['Tables_in_PMA-storage']
+        );
+
+        $this->dummyDbi->addResult(
+            'SELECT NULL FROM `pma__userconfig_custom` LIMIT 0',
+            [
+                ['NULL'],
+            ],
+            ['NULL']
+        );
+
+        $this->dummyDbi->addSelectDb('PMA-storage');
+
+        $relation = new Relation($this->dbi);
+        $relation->initRelationParamsCache();
+
+        $this->assertArrayHasKey(
+            'relation',
+            $_SESSION,
+            'The cache is expected to be filled because the custom override'
+            . 'was undertood (pma__userconfig vs pma__userconfig_custom)'
+        );
+
+        $this->assertAllQueriesConsumed();
+        $this->assertAllSelectsConsumed();
+
+        $this->dummyDbi->addResult(
+            'SHOW TABLES FROM `PMA-storage`',
+            [
+                [
+                    'pma__userconfig_custom',
+                    'pma__usergroups',
+                ],
+            ],
+            ['Tables_in_PMA-storage']
+        );
+
+        $this->dummyDbi->addResult(
+            'SELECT NULL FROM `pma__userconfig_custom` LIMIT 0',
+            [
+                ['NULL'],
+            ],
+            ['NULL']
+        );
+
+        $this->dummyDbi->addSelectDb('PMA-storage');
+        $relationData = (new Relation($this->dbi))->checkRelationsParam();
+        $this->assertAllSelectsConsumed();
+
+        $this->assertSame([
+            'version' => $relationData['version'],
+            'relwork' => false,
+            'displaywork' => false,
+            'bookmarkwork' => false,
+            'pdfwork' => false,
+            'commwork' => false,
+            'mimework' => false,
+            'historywork' => false,
+            'recentwork' => false,
+            'favoritework' => false,
+            'uiprefswork' => false,
+            'trackingwork' => false,
+            'userconfigwork' => true,
+            'menuswork' => false,
+            'navwork' => false,
+            'savedsearcheswork' => false,
+            'centralcolumnswork' => false,
+            'designersettingswork' => false,
+            'exporttemplateswork' => false,
+            'allworks' => false,
+            'user' => '',
+            'db' => 'PMA-storage',
+            'userconfig' => 'pma__userconfig_custom',
+        ], $relationData);
+
+        $this->assertSame([
+            'user' => '',
+            'pmadb' => 'PMA-storage',
+            'bookmarktable' => '',
+            'relation' => '',
+            'table_info' => '',
+            'table_coords' => '',
+            'column_info' => '',
+            'pdf_pages' => '',
+            'history' => '',
+            'recent' => '',
+            'favorite' => '',
+            'table_uiprefs' => '',
+            'tracking' => '',
+            'userconfig' => 'pma__userconfig_custom',
+            'users' => '',
+            'usergroups' => '',
+            'navigationhiding' => '',
+            'savedsearches' => '',
+            'central_columns' => '',
+            'designer_settings' => '',
+            'export_templates' => '',
+        ], $GLOBALS['cfg']['Server']);
+
+        $this->assertAllQueriesConsumed();
+    }
 }
